fix: 회원가입 뷰 SafeArea로 감쌈
